Tests refactoring, added new tests, samples and ethalons, adjusted test files path on macOS
