+++ FileFormat.Heic.Tests/Tests.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * FileFormat.HEIC 
  * Copyright (c) 2024 Openize Pty Ltd. 
  *
@@ -10,35 +10,11 @@
 
 namespace FileFormat.Heic.Tests
 {
-    using NUnit.Framework;
     using FileFormat.Heic.Decoder;
-    using System;
-    using System.Reflection;
-    using System.Runtime.InteropServices;
-    using System.Runtime.Intrinsics;
+    using System.Threading.Tasks;
 
-    public class Tests
+    internal class Tests : TestsCore
     {
-        /// <summary>
-        /// Samples path.
-        /// </summary>
-        private string SamplesPath { get; set; }
-
-        /// <summary>
-        /// Ethalons path.
-        /// </summary>
-        private string EthalonsPath { get; set; }
-
-        /// <summary>
-        /// Startup setup.
-        /// </summary>
-        [SetUp]
-        public void Setup()
-        {
-            SamplesPath = GetSamplesPath();
-            EthalonsPath = GetEthalonsPath();
-        }
-
         /// <summary>
         /// Test decoding of the images generated by iphone.
         /// </summary>
@@ -94,7 +70,7 @@ namespace FileFormat.Heic.Tests
 
         /// <summary>
         /// Test decoding of image with alpha data.
-        /// Image is generated in Gimp.
+        /// Image is created with Gimp.
         /// </summary>
         [Test]
         [TestCase("gimp_rgb_420_with_alpha.heic")]
@@ -141,94 +117,73 @@ namespace FileFormat.Heic.Tests
         }
 
         /// <summary>
-        /// Create ethalon file.
+        /// Test decoding of the rotated and mirrored image.
+        /// Image is captured with iPhone 14 Pro.
         /// </summary>
-        /// <param name="filename">File name.</param>
-        /// <param name="data">Color data.</param>
-        private void CreateEthalon(string filename, byte[] data)
+        [Test]
+        [TestCase("iphone_rotated_and_mirrored.heic")]
+        public void TestRotatedAndMirrored(string filename)
         {
-            var outputFilename = filename + ".bin";
-
-            using (var stream = new FileStream(Path.Combine(EthalonsPath, outputFilename), FileMode.OpenOrCreate, FileAccess.Write))
+            using (var fs = new FileStream(Path.Combine(SamplesPath, filename), FileMode.Open))
             {
-                stream.Write(data, 0, data.Length);
+                var image = HeicImage.Load(fs);
+                var pixels = image.GetByteArray(PixelFormat.Argb32);
+                CompareWithEthalon(filename, pixels);
             }
         }
 
         /// <summary>
-        /// Compare color data with ethalon file.
+        /// Test decoding of the image with log2CbSize issue.
+        /// Image is captured with iPhone 14 Pro.
         /// </summary>
-        /// <param name="filename">File name.</param>
-        /// <param name="data">Color data.</param>
-        private void CompareWithEthalon(string filename, byte[] data)
+        [Test]
+        [TestCase("iphone_log2CbSize_issue.heic")]
+        public void TestLog2CbSizeIssue(string filename)
         {
-            var outputFilename = filename + ".bin";
-
-            using (var stream = new FileStream(Path.Combine(EthalonsPath, outputFilename), FileMode.Open))
+            using (var fs = new FileStream(Path.Combine(SamplesPath, filename), FileMode.Open))
             {
-                const int bytesToRead = 32;
-                int index = 0;
-
-                if (stream.Length != data.Length)
-                {
-                    Assert.Fail($"Ethalon length do not match. Ethalon length equals {stream.Length}, read data length equals {data.Length}");
-                }
-
-                var one = new byte[bytesToRead];
-                var two = new byte[bytesToRead];
-
-                var canRead = true;
-                while (canRead)
-                {
-                    var read_bytes_from_stream = stream.Read(one, 0, bytesToRead);
-
-                    if (index + bytesToRead <= data.Length)
-                        Array.Copy(data, index, two, 0, bytesToRead);
-                    else
-                        Array.Copy(data, index, two, 0, data.Length - index);
-
-                    index += bytesToRead;
-
-                    var vOne = MemoryMarshal.Cast<byte, Vector256<byte>>(one);
-                    var vTwo = MemoryMarshal.Cast<byte, Vector256<byte>>(two);
-
-                    if (!vTwo.SequenceEqual(vOne))
-                    {
-                        Assert.Fail("Data does not match ethalon");
-                        return;
-                    }
-
-                    canRead = read_bytes_from_stream == bytesToRead && index < data.Length;
-                }
-
-                Assert.Pass();
+                var image = HeicImage.Load(fs);
+                var pixels = image.GetByteArray(PixelFormat.Argb32);
+                CompareWithEthalon(filename, pixels);
             }
         }
 
         /// <summary>
-        /// Get project path.
+        /// Test decoding of the image with CleanAperture Box.
+        /// Image is created with Gimp.
         /// </summary>
-        private static string GetProjectPath()
+        [Test]
+        [TestCase("gimp_with_clap.heic")]
+        public void TestCleanApertureBox(string filename)
         {
-            var path = new Uri(Assembly.GetExecutingAssembly().GetName().CodeBase).LocalPath;
-            return path.Remove(path.IndexOf("\\bin"));
+            using (var fs = new FileStream(Path.Combine(SamplesPath, filename), FileMode.Open))
+            {
+                var image = HeicImage.Load(fs);
+                var pixels = image.GetByteArray(PixelFormat.Argb32);
+                CompareWithEthalon(filename, pixels);
+            }
         }
 
         /// <summary>
-        /// Get test samples path.
+        /// Test decoding of the parallel processing decoding.
         /// </summary>
-        private static string GetSamplesPath()
+        [Test]
+        public void TestParallelProcessing()
         {
-            return Path.Combine(GetProjectPath(), "TestsData", "samples");
-        }
+            Task t1 = Task.Run(() => Read(Path.Combine(SamplesPath, "nokia", "grid_960x640.HEIC")));
+            Task t2 = Task.Run(() => Read(Path.Combine(SamplesPath, "nokia", "overlay_1000x680.HEIC")));
+            Task t3 = Task.Run(() => Read(Path.Combine(SamplesPath, "nokia", "random_collection_1440x960.HEIC")));
 
-        /// <summary>
-        /// Get test ethalons path.
-        /// </summary>
-        private static string GetEthalonsPath()
-        {
-            return Path.Combine(GetProjectPath(), "TestsData", "ethalons");
-        }
+            Task.WaitAll(t1, t2, t3);
 
+            void Read(string filename)
+            {
+                using (var fs = new FileStream(filename, FileMode.Open))
+                {
+                    var image = HeicImage.Load(fs);
+                    var pixels = image.GetInt32Array(PixelFormat.Argb32);
+                }
+            }
+        }
     }
 }
